move file and prepare registery pattern archi
