+++ .gitignore
@@ -32,3 +32,7 @@ Thumbs.db
 # Your specific files (add as needed)
 # logs/
 # temp/
+
+# coverage
+.coverage
+
